+++ custom_components/eltako/schema.py
@@ -150,6 +150,7 @@ class SensorSchema(EltakoPlatformSchema):
     CONF_EEP_SUPPORTED = [A5_04_01.eep_string,
                           A5_04_02.eep_string,
                           A5_07_01.eep_string,
+                          A5_08_01.eep_string,
                           A5_09_0C.eep_string,
                           A5_10_06.eep_string,
                           A5_10_12.eep_string,